Some Fixes

+++ src/download.ts
@@ -4,9 +4,9 @@ import ffmpeg from 'ffmpeg-static';
 import fs from 'fs'
 import { Readable, Writable } from 'stream';
 
-let fileIndexes: number[] = [];
+let fileIndexes: string[] = [];
 
-type videoFile = {fileIndex: number, fileName: string, finalReadable: Readable};
+type videoFile = { fileName: string | null, finalReadable: Readable };
 type formats = 'mp4' | 'webm';
 
 /**
@@ -14,10 +14,12 @@ type formats = 'mp4' | 'webm';
  * @param index file index
  * @param name file name
  */
-function deleteFile(index: number, name: string) {
-    fs.unlink(`./downloads/${name}`, () => {});
-    fileIndexes.splice(fileIndexes.indexOf(index), 1);
+function deleteFile( name: string | null) {
+    if (!name) return;
+    fs.unlink(`./downloads/${name}`, () => {}); // Didn't really know how to manage temporary stored videos so I did this.
+    fileIndexes.splice(fileIndexes.indexOf(name), 1);
 }
+
 /**
  * Merges video and audio streams.
  * @param video Readable video stream
@@ -25,7 +27,7 @@ function deleteFile(index: number, name: string) {
  * @param title Video Title
  * @returns ffmpeg Child Process
  */
-function mergeAudioAndVideo(video: Readable, audio: Readable, title: string) {
+function mergeAudioAndVideo(video: Readable, audio: Readable, fileName: string) {
     const ffmpegProcess = cp.spawn(ffmpeg as string, [
         '-i', `pipe:3`,
         '-i', `pipe:4`,
@@ -36,7 +38,7 @@ function mergeAudioAndVideo(video: Readable, audio: Readable, title: string) {
         '-crf', '17',
         '-preset', 'veryfast',
         '-f', 'mp4',
-        `./downloads/${title}`
+        `./downloads/${fileName}`
     ], {
         stdio: [
             'pipe', 'pipe', 'pipe', 'pipe', 'pipe',
@@ -84,16 +86,63 @@ async function getMetadata(url: string) {
 }
 
 /**
- * 
+ * Downloads both video and audio streams and merges them together using ffmpeg. Only works with mp4 for the moment.
+ * It does this by creating a temp file where ffmpeg merges video and audio streams (This file then has to be deleted).
  * @param url video url
  * @param itag valid youtube itag
  * @param format can be mp4, webm or mp3
  * @returns VideoFile type with readable video stream
  */
-async function downloadVideo(url: string, itag: number, format: formats) {
+async function downloadMergedVideo(url: string, itag: number, format: formats) {
     let downloadProgress: any;
 
-    const videoStream = ytdl(url, { quality: itag }).on('error', (err) => {
+    const videoStream = ytdl(url, { quality: itag, highWaterMark: 1 << 25 }).on('error', (err) => {
+        console.log(err);
+        return;
+    }).on('progress', (length, downloaded, totallength) => {
+        // Calculate download progress.
+        const downloadedProgress = Math.round(downloaded * 100 / totallength);
+        if (downloadProgress !== downloadedProgress) {
+            downloadProgress = downloadedProgress;
+            console.clear();
+            const completion = progressBar(downloadProgress);
+            console.log(completion);
+        }
+    });
+    
+    const audioStream = ytdl(url, { filter: 'audioonly', quality: 'highestaudio', highWaterMark: 1 << 25 });
+        
+    // Creates a temprary file to merge video and audio, which then gets sent to the user and deleted.
+    let fileNum = 1;
+    let fileName = `temp${fileNum}.mp4`;
+    while (fileIndexes.includes(fileName) || fs.existsSync(`./downloads/${fileName}`)) {
+        if(!fileIndexes.includes(fileName) && fs.existsSync(`./downloads/${fileName}`)) {
+            fileIndexes.push(fileName);
+        } 
+        fileNum++;
+        fileName = `temp${fileNum}.mp4`;
+    }
+    fileIndexes.push(fileName);
+    
+    await new Promise<void>((resolve) => {
+        mergeAudioAndVideo(videoStream, audioStream, fileName)
+        .on('close', () => {
+          resolve();
+        })
+    }) 
+
+    let final: Readable = videoStream;
+    audioStream.destroy();
+    videoStream.destroy();
+    final = fs.createReadStream(`./downloads/${fileName}`);
+
+    return { fileName: fileName, finalReadable: final} as videoFile;
+}
+
+function downloadVideo(url: string, itag: number) {
+    let downloadProgress: any;
+    
+    const videoStream = ytdl(url, { quality: itag, highWaterMark: 1 << 25 }).on('error', (err) => {
         console.log(err);
         return;
     }).on('progress', (length, downloaded, totallength) => {
@@ -107,33 +156,7 @@ async function downloadVideo(url: string, itag: number, format: formats) {
         }
     });
 
-    let final: Readable = videoStream;
-
-    let fileNum = 1;
-    let fileName = `temp${fileNum}.mp4`;
-    if (itag !== 18 && itag !== 22 && format !== 'webm') {
-        const audioStream = ytdl(url, { filter: 'audioonly', quality: 'highestaudio' });
-        while (fileIndexes.includes(fileNum) || fs.existsSync(`./downloads/${fileName}`)) {
-            if(!fileIndexes.includes(fileNum)) {
-                fileIndexes.push(fileNum);
-            } else if (fileIndexes.includes(fileNum) && !fs.existsSync(`./downloads/${fileName}`)) {
-                fileIndexes.splice(fileIndexes.indexOf(fileNum), 1);
-            }
-            fileNum++;
-            fileName = `temp${fileNum}.mp4`;
-        }
-        fileIndexes.push(fileNum);
-        
-        await new Promise<void>((resolve) => { // wait
-            mergeAudioAndVideo(videoStream, audioStream, fileName)
-            .on('close', () => {
-              resolve(); // finish
-            })
-        })
-        final = fs.createReadStream(`./downloads/${fileName}`);
-    }
-
-    return { fileIndex: fileNum, fileName: fileName, finalReadable: final} as videoFile;
+    return videoStream;
 }
 
 /**
@@ -168,17 +191,15 @@ function progressBar(progress: number) {
             progressBar += ' ';
         }
     }
-    progressBar += ` ] ${progress}%`;
+    progressBar += `] ${progress}%`;
 
     return progressBar;
 }
 
-// test
-// downloadVideo('https://www.youtube.com/watch?v=QjCkFem2y0U', 137);
-
 export {
     getMetadata,
     downloadVideo,
+    downloadMergedVideo,
     downloadAudioOnly,
     deleteFile,
     formats,
